+++ src/lib.rs
@@ -6,8 +6,14 @@
 //! This driver allows you to:
 //! - Enable/disable the device.
 //! - Read the temperature.
+//! - Set the fault queue.
+//! - Set the OS temperature.
+//! - Set the hysteresis temperature.
+//! - Set the OS operation mode.
+//! - Set the OS polarity.
 //!
 //! ## The device
+//!
 //! The LM75 temperature sensor includes a delta-sigma analog-to-digital
 //! converter, and a digital overtemperature detector. The host can
 //! query the LM75 through its I2C interface to read temperature at any
@@ -74,6 +80,89 @@
 //! # }
 //! ```
 //!
+//! ### Set the fault queue
+//!
+//! This is the number of consecutive faults necessary to trigger
+//! an OS condition.
+//!
+//! ```no_run
+//! extern crate linux_embedded_hal as hal;
+//! extern crate lm75;
+//!
+//! use hal::I2cdev;
+//! use lm75::{ Lm75, SlaveAddr, FaultQueue };
+//!
+//! # fn main() {
+//! let dev = I2cdev::new("/dev/i2c-1").unwrap();
+//! let mut sensor = Lm75::new(dev, SlaveAddr::default());
+//! sensor.set_fault_queue(FaultQueue::_4).unwrap();
+//! # }
+//! ```
+//!
+//! ### Set the OS polarity
+//!
+//! ```no_run
+//! extern crate linux_embedded_hal as hal;
+//! extern crate lm75;
+//!
+//! use hal::I2cdev;
+//! use lm75::{ Lm75, SlaveAddr, OsPolarity };
+//!
+//! # fn main() {
+//! let dev = I2cdev::new("/dev/i2c-1").unwrap();
+//! let mut sensor = Lm75::new(dev, SlaveAddr::default());
+//! sensor.set_os_polarity(OsPolarity::ActiveHigh).unwrap();
+//! # }
+//! ```
+//!
+//! ### Set the OS operation mode
+//!
+//! ```no_run
+//! extern crate linux_embedded_hal as hal;
+//! extern crate lm75;
+//!
+//! use hal::I2cdev;
+//! use lm75::{ Lm75, SlaveAddr, OsMode };
+//!
+//! # fn main() {
+//! let dev = I2cdev::new("/dev/i2c-1").unwrap();
+//! let mut sensor = Lm75::new(dev, SlaveAddr::default());
+//! sensor.set_os_mode(OsMode::Interrupt).unwrap();
+//! # }
+//! ```
+//!
+//! ### Set the OS temperature
+//!
+//! ```no_run
+//! extern crate linux_embedded_hal as hal;
+//! extern crate lm75;
+//!
+//! use hal::I2cdev;
+//! use lm75::{ Lm75, SlaveAddr };
+//!
+//! # fn main() {
+//! let dev = I2cdev::new("/dev/i2c-1").unwrap();
+//! let mut sensor = Lm75::new(dev, SlaveAddr::default());
+//! sensor.set_os_temperature(50.0).unwrap();
+//! # }
+//! ```
+//!
+//! ### Set the hysteresis temperature
+//!
+//! ```no_run
+//! extern crate linux_embedded_hal as hal;
+//! extern crate lm75;
+//!
+//! use hal::I2cdev;
+//! use lm75::{ Lm75, SlaveAddr };
+//!
+//! # fn main() {
+//! let dev = I2cdev::new("/dev/i2c-1").unwrap();
+//! let mut sensor = Lm75::new(dev, SlaveAddr::default());
+//! sensor.set_hysteresis_temperature(40.0).unwrap();
+//! # }
+//! ```
+//!
 //! ### Enable / disable the sensor
 //!
 //! ```no_run
@@ -103,6 +192,8 @@ use hal::blocking::i2c;
 pub enum Error<E> {
     /// I²C bus error
     I2C(E),
+    /// Invalid input data
+    InvalidInputData,
 }
 
 /// Possible slave addresses
@@ -133,6 +224,39 @@ impl SlaveAddr {
     }
 }
 
+/// Fault queue
+///
+/// Number of consecutive faults necessary to trigger OS condition.
+#[derive(Debug, Clone)]
+pub enum FaultQueue {
+    /// 1 fault will trigger OS condition
+    _1,
+    /// 2 consecutive faults will trigger OS condition
+    _2,
+    /// 4 consecutive faults will trigger OS condition
+    _4,
+    /// 6 consecutive faults will trigger OS condition
+    _6,
+}
+
+/// OS polarity
+#[derive(Debug, Clone)]
+pub enum OsPolarity {
+    /// Active low
+    ActiveLow,
+    /// Active high
+    ActiveHigh
+}
+
+/// OS operation mode
+#[derive(Debug, Clone)]
+pub enum OsMode {
+    /// Comparator
+    Comparator,
+    /// Interrupt
+    Interrupt
+}
+
 const DEVICE_BASE_ADDRESS: u8 = 0b100_1000;
 
 struct Register;
@@ -140,6 +264,8 @@ struct Register;
 impl Register {
     const TEMPERATURE   : u8 = 0x00;
     const CONFIGURATION : u8 = 0x01;
+    const T_HYST        : u8 = 0x02;
+    const T_OS          : u8 = 0x03;
 }
 
 
@@ -147,6 +273,10 @@ struct BitFlags;
 
 impl BitFlags {
     const SHUTDOWN     : u8 = 0b0000_0001;
+    const COMP_INT     : u8 = 0b0000_0010;
+    const OS_POLARITY  : u8 = 0b0000_0100;
+    const FAULT_QUEUE0 : u8 = 0b0000_1000;
+    const FAULT_QUEUE1 : u8 = 0b0001_0000;
 }
 
 /// LM75 device driver.
@@ -190,6 +320,62 @@ where
     pub fn disable(&mut self) -> Result<(), Error<E>> {
         let config = self.config;
         self.write_config(config | BitFlags::SHUTDOWN)
+    }
+
+    /// Set the fault queue.
+    ///
+    /// Set the number of consecutive faults that will trigger an OS condition.
+    pub fn set_fault_queue(&mut self, fq: FaultQueue) -> Result<(), Error<E>> {
+        let mut config = self.config;
+        match fq {
+            FaultQueue::_1 => config = config & !BitFlags::FAULT_QUEUE1 & !BitFlags::FAULT_QUEUE0,
+            FaultQueue::_2 => config = config & !BitFlags::FAULT_QUEUE1 |  BitFlags::FAULT_QUEUE0,
+            FaultQueue::_4 => config = config |  BitFlags::FAULT_QUEUE1 & !BitFlags::FAULT_QUEUE0,
+            FaultQueue::_6 => config = config |  BitFlags::FAULT_QUEUE1 |  BitFlags::FAULT_QUEUE0,
+        }
+        self.write_config(config)
+    }
+
+    /// Set the OS polarity.
+    pub fn set_os_polarity(&mut self, polarity: OsPolarity) -> Result<(), Error<E>> {
+        let mut config = self.config;
+        match polarity {
+            OsPolarity::ActiveLow  => config = config & !BitFlags::OS_POLARITY,
+            OsPolarity::ActiveHigh => config = config |  BitFlags::OS_POLARITY,
+        }
+        self.write_config(config)
+    }
+
+    /// Set the OS operation mode.
+    pub fn set_os_mode(&mut self, mode: OsMode) -> Result<(), Error<E>> {
+        let mut config = self.config;
+        match mode {
+            OsMode::Comparator => config = config & !BitFlags::COMP_INT,
+            OsMode::Interrupt  => config = config |  BitFlags::COMP_INT,
+        }
+        self.write_config(config)
+    }
+
+    /// Set the OS temperature.
+    pub fn set_os_temperature(&mut self, temperature: f32) -> Result<(), Error<E>> {
+        if temperature < -55.0 || temperature > 125.0 {
+            return Err(Error::InvalidInputData);
+        }
+        let (msb, lsb) = conversion::convert_temp_to_register(temperature);
+        self.i2c
+            .write(self.address, &[Register::T_OS, msb, lsb])
+            .map_err(Error::I2C)
+    }
+
+    /// Set the hysteresis temperature.
+    pub fn set_hysteresis_temperature(&mut self, temperature: f32) -> Result<(), Error<E>> {
+        if temperature < -55.0 || temperature > 125.0 {
+            return Err(Error::InvalidInputData);
+        }
+        let (msb, lsb) = conversion::convert_temp_to_register(temperature);
+        self.i2c
+            .write(self.address, &[Register::T_HYST, msb, lsb])
+            .map_err(Error::I2C)
     }
 
     fn write_config(&mut self, config: u8) -> Result<(), Error<E>> {